Move RenderBuffer back and add Mask and MaskManager
