+++ source/Body.cpp
@@ -18,8 +18,8 @@ this program. If not, see <https://www.gnu.org/licenses/>.
 #include "DataNode.h"
 #include "DataWriter.h"
 #include "GameData.h"
-#include "Mask.h"
-#include "MaskManager.h"
+#include "images/Mask.h"
+#include "images/MaskManager.h"
 #include "pi.h"
 #include "Random.h"
 #include "image/Sprite.h"
