+++ source/GameData.cpp
@@ -38,7 +38,7 @@ this program. If not, see <https://www.gnu.org/licenses/>.
 #include "image/ImageSet.h"
 #include "Interface.h"
 #include "LineShader.h"
-#include "MaskManager.h"
+#include "images/MaskManager.h"
 #include "Minable.h"
 #include "Mission.h"
 #include "audio/Music.h"
@@ -51,7 +51,7 @@ this program. If not, see <https://www.gnu.org/licenses/>.
 #include "Plugins.h"
 #include "PointerShader.h"
 #include "Politics.h"
-#include "image/RenderBuffer.h"
+#include "RenderBuffer.h"
 #include "RingShader.h"
 #include "Ship.h"
 #include "image/Sprite.h"
