+++ source/PreferencesPanel.cpp
@@ -29,7 +29,7 @@ this program. If not, see <https://www.gnu.org/licenses/>.
 #include "Plugins.h"
 #include "PointerShader.h"
 #include "Preferences.h"
-#include "image/RenderBuffer.h"
+#include "RenderBuffer.h"
 #include "Screen.h"
 #include "image/Sprite.h"
 #include "image/SpriteSet.h"
