+++ source/RenderBuffer.h
@@ -15,8 +15,8 @@ this program. If not, see <https://www.gnu.org/licenses/>.
 
 #pragma once
 
-#include "../Point.h"
-#include "../Screen.h"
+#include "Point.h"
+#include "Screen.h"
