+++ source/images/Mask.cpp
@@ -15,8 +15,8 @@ this program. If not, see <https://www.gnu.org/licenses/>.
 
 #include "Mask.h"
 
-#include "image/ImageBuffer.h"
-#include "Logger.h"
+#include "../image/ImageBuffer.h"
+#include "../Logger.h"
 
 #include <algorithm>
 #include <cmath>
