+++ source/images/Mask.h
@@ -15,8 +15,8 @@ this program. If not, see <https://www.gnu.org/licenses/>.
 
 #pragma once
 
-#include "Angle.h"
-#include "Point.h"
+#include "../Angle.h"
+#include "../Point.h"
 
 #include <vector>
