+++ source/images/MaskManager.cpp
@@ -15,8 +15,8 @@ this program. If not, see <https://www.gnu.org/licenses/>.
 
 #include "MaskManager.h"
 
-#include "Logger.h"
-#include "image/Sprite.h"
+#include "../Logger.h"
+#include "../image/Sprite.h"
 
 using namespace std;
