+++ source/images/MaskManager.h
